Refactored the code, so that easily add objects. Sheep is white with black eyes.
